requirements.txt
fastmcp
langchain
python-dotenv
langchain_community
langchain_core
requests
tavily-python